varlist: move to a separate header, add join
